os-3382 Move HAPROXY ServiceContainer creation to when HM is created
Don't create the HAPROXY ServiceContainer and its Port when the LoadBalancer
is created but wait for the HealthMonitor. The service container won't start
as long as there's no HealthMonitor. And if a LoadBalancer is set up without
a HealthMonitor the service container will never start. Trying to schedule
it to some host doesn't make sense. It times out and is re-scheduled to a
different host over and over again. That in turn creates a lot of
Zookeeper watchers in midolman (Port host_id changes, midolman updates
observable on /midonet/zoom/0/state/<host-id>/Port/<port-id>/active,
old watcher is closed, new watcher is started). But the old watcher isn't
really removed from Zookeeper, only partly in Curator --> we have a leak.

With the changes here the ServiceContainerGroup, ServiceContainer and Port
for the health monitoring (haproxy) are only created when the HealthMonitor
is created.