RuleProtocol: Support for all IP protocols (#26)
Security group rules have an optional "protocol" field to let the
rule match only for a given IP protocol (e.g. "udp" or "17").
Internally Midonet implemented the allowed protocols in an enum
and if the user configures a protocol not covered by the enum, the
protocol condition is ignored. Without the protocol condition too
many packets may pass.
E.g. if UDP is restricted to a certain port by one rule and IGMP is
allowed by another one (ingress, ethertype IPv4, protocol "igmp")
the 2nd rule is saved by Midonet as (ingress, ethertype IPv4) which
will allow all UDP/TCP.

A previous change added a few protocols to the enum.
Here we complete the list of protocols, so the protocol attribute
of a rule is never ignored (unless an unknown string is passed,
but that should be handled by Neutron).